Updated code accordingly to the PR review  comments
