+++ page_objects/editSpecialtyPage.ts
@@ -1,4 +1,4 @@
-import { Page, expect } from '@playwright/test';
+import { Page } from '@playwright/test';
 
 export class EditSpecialtyPage {
     readonly page: Page
