+++ page_objects/specialtiesPage.ts
@@ -1,9 +1,9 @@
 import { Page, expect } from '@playwright/test';
-import { HelperBase } from '../page_objects/helperBase';
 
-export class SpecialtiesPage extends HelperBase{
+export class SpecialtiesPage {
+    readonly page: Page
     constructor(page: Page){
-        super(page)
+        this.page = page
     }
 
     async deleteSpecialtyByName(specialty: string){
